feat: add na override

+++ build.sh
@@ -27,8 +27,14 @@ then
 fi
 
 clean () {
+  #dont need to exec twice
+  if [ "$FIRST_TIME" = "false" ]; then
+    return 0
+  fi
   rm -rf node_modules
   rm -rf dist
+  rm -rf build
+  rm -rf deployment/*
 }
 
 init () {
@@ -38,21 +44,34 @@ init () {
     echo "[init] Get branch name from git..."
     BRANCH_NAME=`git branch | sed -n -e "s/^\* \(.*\)/\1/p"`
   fi
-  docker-compose run -e BRANCH_NAME=$BRANCH_NAME -e FRONT_TAG=$FRONT_TAG -e NEXUS_ODE_USERNAME=$NEXUS_ODE_USERNAME -e NEXUS_ODE_PASSWORD=$NEXUS_ODE_PASSWORD --rm -u "$USER_UID:$GROUP_GID" gradle sh -c "gradle generateTemplate"
+  docker-compose run -e OVERRIDE_NAME=$OVERRIDE_NAME -e OVERRIDE_MODNAME=$OVERRIDE_MODNAME -e BRANCH_NAME=$BRANCH_NAME -e FRONT_TAG=$FRONT_TAG -e NEXUS_ODE_USERNAME=$NEXUS_ODE_USERNAME -e NEXUS_ODE_PASSWORD=$NEXUS_ODE_PASSWORD --rm -u "$USER_UID:$GROUP_GID" gradle sh -c "gradle generateTemplate"
+  #dont need to exec twice
+  if [ "$FIRST_TIME" = "false" ]; then
+    return 0
+  fi
   docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm rebuild node-sass --no-bin-links && npm install"
 }
 
 build () {
   local extras=$1
+  #get skins
   dirs=($(ls -d ./skins/*))
-  docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm run release:prepare"
+  #create build dir var
+  SCSS_DIR=$OVERRIDE_BUILD/scss
+  SKIN_DIR=$OVERRIDE_BUILD/skins
+  #create build dir and copy source
+  mkdir -p $OVERRIDE_BUILD
+  cp -R skins $SKIN_DIR
+  cp -R scss $SCSS_DIR
+  cp -R scss/$OVERRIDE_SRC/* $OVERRIDE_BUILD/scss/
+  docker-compose run -e SKIN_DIR=$SKIN_DIR -e SCSS_DIR=$SCSS_DIR -e DIST_DIR=$OVERRIDE_DIST --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm run release:prepare"
   for dir in "${dirs[@]}"; do
     tmp=`echo $dir | sed 's/.\/skins\///'`
-    docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "SKIN=$tmp npm run sass:build:release"
+    docker-compose run -e SKIN_DIR=$SKIN_DIR -e SCSS_DIR=$SCSS_DIR -e DIST_DIR=$OVERRIDE_DIST -e SKIN=$tmp  --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm run sass:build:release"
   done
-  cp node_modules/ode-bootstrap/version.txt dist/version.txt
+  cp node_modules/ode-bootstrap/version.txt $OVERRIDE_DIST/version.txt
   VERSION=`grep "version="  gradle.properties| sed 's/version=//g'`
-  echo "ode-bootstrap-neo=$VERSION `date +'%d/%m/%Y %H:%M:%S'`" >> dist/version.txt
+  echo "ode-bootstrap-neo=$VERSION `date +'%d/%m/%Y %H:%M:%S'`" >> $OVERRIDE_DIST/version.txt
 }
 
 watch () {
@@ -60,6 +79,10 @@ watch () {
 }
 
 publishNPM () {
+  #dont need to exec twice
+  if [ "$FIRST_TIME" = "false" ]; then
+    return 0
+  fi
   LOCAL_BRANCH=`echo $GIT_BRANCH | sed -e "s|origin/||g"`
   docker-compose run --rm -u "$USER_UID:$GROUP_GID" node sh -c "npm publish --tag $LOCAL_BRANCH"
 }
@@ -72,41 +95,63 @@ publishNexus () {
     echo "sonatypeUsername=$NEXUS_SONATYPE_USERNAME" >> "?/.gradle/gradle.properties"
     echo "sonatypePassword=$NEXUS_SONATYPE_PASSWORD" >> "?/.gradle/gradle.properties"
   fi
-  docker-compose run --rm -u "$USER_UID:$GROUP_GID" gradle sh -c "gradle deploymentJar fatJar publish"
+  docker-compose run -e OVERRIDE_NAME=$OVERRIDE_NAME -e OVERRIDE_MODNAME=$OVERRIDE_MODNAME --rm -u "$USER_UID:$GROUP_GID" gradle sh -c "gradle deploymentJar fatJar publish"
 }
 
 publishMavenLocal(){
-  docker-compose run --rm -u "$USER_UID:$GROUP_GID" gradle sh -c "gradle deploymentJar fatJar publishToMavenLocal"
+  docker-compose run -e OVERRIDE_NAME=$OVERRIDE_NAME -e OVERRIDE_MODNAME=$OVERRIDE_MODNAME --rm -u "$USER_UID:$GROUP_GID" gradle sh -c "gradle deploymentJar fatJar publishToMavenLocal"
 }
 
+
+export FIRST_TIME=true
 for param in "$@"
 do
-  case $param in
-    clean)
-      clean
-      ;;
-    init)
-      init
-      ;;
-    build)
-      build
-      ;;
-    install)
-      build && publishMavenLocal
-      ;;
-    watch)
-      watch
-      ;;
-    publishNPM)
-      publishNPM
-      ;;
-    publishNexus)
-      publishNexus
-      ;;
-    *)
-      echo "Invalid argument : $param"
-  esac
-  if [ ! $? -eq 0 ]; then
-    exit 1
-  fi
+  names=($(ls -d ./scss/overrides/*))
+  for dirtyName in "${names[@]}"; do
+    name=`echo $dirtyName | sed 's/.\/scss\/overrides\///'`
+    MOD_NAME=`grep "modname="  gradle.properties| sed 's/modname=//g'`
+    export OVERRIDE_NAME="$name"
+    export OVERRIDE_SRC="overrides/$name"
+    export OVERRIDE_DIST="dist/$name"
+    export OVERRIDE_DIST_WILDCARD="dist/$name/**"
+    export OVERRIDE_BUILD="build/$name"
+    export OVERRIDE_MODNAME="$MOD_NAME-$name"
+    #default does not have suffix on name
+    if [ "$name" = "default" ]; then
+      export OVERRIDE_MODNAME="$MOD_NAME"
+    fi
+    echo "[$param][$name] Starting..."
+    case $param in
+      clean)
+        clean
+        ;;
+      init)
+        init
+        ;;
+      build)
+        build
+        ;;
+      install)
+        build && publishMavenLocal
+        ;;
+      publishMavenLocal)
+        publishMavenLocal
+        ;;
+      watch)
+        watch
+        ;;
+      publishNPM)
+        publishNPM
+        ;;
+      publishNexus)
+        publishNexus
+        ;;
+      *)
+        echo "Invalid argument : $param"
+    esac
+    if [ ! $? -eq 0 ]; then
+      exit 1
+    fi
+    export FIRST_TIME=false
+  done
 done
